refactor(mux): use Chains enum
